requirement.txt
langchain-google-genai
langchain
streamlit
pandas 
numpy
python-dotenv
tk
ensure==1.0.2
mlflow==2.2.2
tensorflow==2.12.0
dvc
opencv-python
python-box==6.0.2
gdown
joblib
scikit-learn
google-generativeai
keras
xgboost
streamlit_option_menu
pytesseract
pillow
PyPDF2
faiss-cpu
-e .